evo/requirements.txt
docling
groq
python-dotenv
